rename(model,view,validation): 폴더 구조화
