+++ src/App.js
@@ -1,9 +1,9 @@
 const { Console } = require("@woowacourse/mission-utils");
 const { GAME_MESSAGE, PLAYER_CHOICE } = require("./utils/Constants");
-const InputView = require("./InputView");
-const OutputView = require("./OutputView");
+const InputView = require("./view/InputView");
+const OutputView = require("./view/OutputView");
 const Bridge = require("./Bridge");
-const BridgeGame = require("./BridgeGame");
+const BridgeGame = require("./model/BridgeGame");
 
 class App {
   #bridge;
