+++ src/view/OutputView.js
@@ -1,5 +1,5 @@
 const { Console } = require("@woowacourse/mission-utils");
-const { GAME_MESSAGE } = require("./utils/Constants");
+const { GAME_MESSAGE } = require("../utils/Constants");
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
  */
